+++ crates/enterprise/unleash-edge-context-enrichers/src/lib.rs
@@ -1,2 +1,5 @@
+mod child;
+mod command;
+mod driver;
 mod protocol;
 mod worker;
